+++ src/components/core/Flyout/Flyout.tsx
@@ -15,32 +15,40 @@ import {
   PrimitiveFlyoutTrigger,
 } from "components/primitives";
 import { flyout } from "generated/panda/recipes";
-import { useIsMounted } from "lib/hooks";
+import { useIsClient } from "lib/hooks";
 
 import type {
   PrimitiveFlyoutProps,
   PrimitiveFlyoutTriggerProps,
 } from "components/primitives";
-import type { ReactNode } from "react";
+import type { RefObject, ReactNode } from "react";
 
 export interface Props extends PrimitiveFlyoutProps {
   trigger?: ReactNode;
   title?: ReactNode;
   children: ReactNode;
   triggerProps?: PrimitiveFlyoutTriggerProps;
+  targetRef?: RefObject<HTMLElement>;
 }
 
 /**
  * Core UI flyout.
  */
-const Flyout = ({ trigger, title, children, triggerProps, ...rest }: Props) => {
+const Flyout = ({
+  trigger,
+  title,
+  children,
+  triggerProps,
+  targetRef,
+  ...rest
+}: Props) => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
 
   const classNames = flyout();
 
-  const isMounted = useIsMounted();
+  const isClient = useIsClient();
 
-  if (!isMounted) return null;
+  if (!isClient) return null;
 
   return (
     <PrimitiveFlyout
@@ -59,7 +67,7 @@ const Flyout = ({ trigger, title, children, triggerProps, ...rest }: Props) => {
         </PrimitiveFlyoutTrigger>
       )}
 
-      <Portal>
+      <Portal target={targetRef}>
         <PrimitiveFlyoutPositioner className={classNames.positioner}>
           <PrimitiveFlyoutContent className={classNames.content}>
             <PrimitiveFlyoutArrow className={classNames.arrow}>