+++ src/components/core/Menu/Menu.tsx
@@ -13,14 +13,14 @@ import {
 } from "components/primitives";
 import { cx } from "generated/panda/css";
 import { button, menu } from "generated/panda/recipes";
-import { useIsMounted } from "lib/hooks";
+import { useIsClient } from "lib/hooks";
 
 import type { PrimitiveMenuProps } from "components/primitives";
 import type {
   ButtonVariantProps,
   MenuVariantProps,
 } from "generated/panda/recipes";
-import type { ReactElement, ReactNode } from "react";
+import type { ReactElement, ReactNode, RefObject } from "react";
 
 export interface MenuItemRecord {
   id: string;
@@ -40,6 +40,7 @@ export interface Props extends PrimitiveMenuProps, MenuVariantProps {
   triggerItem?: ReactNode;
   triggerVariant?: ButtonVariantProps["variant"];
   groups?: MenuItemGroupRecord[];
+  targetRef?: RefObject<HTMLElement>;
 }
 
 /**
@@ -52,13 +53,14 @@ const Menu = ({
   triggerVariant,
   groups,
   size,
+  targetRef,
   ...rest
 }: Props) => {
   const classNames = menu({ size });
 
-  const isMounted = useIsMounted();
+  const isClient = useIsClient();
 
-  if (!isMounted) return null;
+  if (!isClient) return null;
 
   return (
     <PrimitiveMenu {...rest}>
@@ -79,7 +81,7 @@ const Menu = ({
               {triggerItem}
             </PrimitiveMenuTriggerItem>
           )}
-          <Portal>
+          <Portal target={targetRef}>
             <PrimitiveMenuPositioner className={classNames.positioner}>
               <PrimitiveMenuContent className={classNames.content}>
                 {groups?.map(({ id, label, separator, items }) => (