+++ src/components/core/Tooltip/Tooltip.tsx
@@ -9,22 +9,19 @@ import {
   PrimitiveTooltipTrigger,
 } from "components/primitives";
 import { tooltip } from "generated/panda/recipes";
-import { useIsMounted } from "lib/hooks";
+import { useIsClient } from "lib/hooks";
 
-import type {
-  PrimitiveTooltipProps,
-  PrimitiveTooltipTriggerProps,
-} from "components/primitives";
+import type { PrimitiveTooltipProps } from "components/primitives";
 import type { TooltipVariantProps } from "generated/panda/recipes";
 import type { JsxStyleProps } from "generated/panda/types";
-import type { ReactNode } from "react";
+import type { ReactNode, RefObject } from "react";
 
 export interface Props extends PrimitiveTooltipProps, TooltipVariantProps {
   trigger?: ReactNode;
-  content: ReactNode;
+  tooltipContent: ReactNode;
   bgColor?: JsxStyleProps["bgColor"];
-  triggerProps?: PrimitiveTooltipTriggerProps;
   arrow?: boolean;
+  targetRef?: RefObject<HTMLElement>;
 }
 
 /**
@@ -32,36 +29,32 @@ export interface Props extends PrimitiveTooltipProps, TooltipVariantProps {
  */
 const Tooltip = ({
   trigger,
-  content,
+  tooltipContent,
   openDelay = 0,
   closeDelay = 0,
   bgColor = "bg.default",
   variant,
-  triggerProps,
   arrow = true,
+  targetRef,
   ...rest
 }: Props) => {
-  const isMounted = useIsMounted();
+  const isClient = useIsClient();
 
   const classNames = tooltip({ variant });
 
-  if (!isMounted) return null;
+  if (!isClient) return null;
 
   return (
     <PrimitiveTooltip openDelay={openDelay} closeDelay={closeDelay} {...rest}>
       {({ isOpen }) => (
         <>
           {trigger && (
-            <PrimitiveTooltipTrigger
-              asChild
-              className={classNames.trigger}
-              {...triggerProps}
-            >
+            <PrimitiveTooltipTrigger asChild className={classNames.trigger}>
               {trigger}
             </PrimitiveTooltipTrigger>
           )}
 
-          <Portal>
+          <Portal target={targetRef}>
             <PrimitiveTooltipPositioner className={classNames.positioner}>
               {isOpen && (
                 <>
@@ -79,7 +72,7 @@ const Tooltip = ({
                     bgColor={bgColor}
                     className={classNames.content}
                   >
-                    {content}
+                    {tooltipContent}
                   </PrimitiveTooltipContent>
                 </>
               )}