+++ src/components/primitives/Accordion/Accordion.tsx
@@ -7,28 +7,29 @@ import {
 
 import { panda } from "generated/panda/jsx";
 
-import type {
-  AccordionProps,
-  AccordionItemProps,
-  AccordionTriggerProps,
-  AccordionContentProps,
-} from "@ark-ui/react";
 import type { PandaComponent } from "generated/panda/types/jsx";
+import type { ComponentProps } from "react";
 
 /**
  * Core UI accordion primitives.
  */
-export type PrimitiveAccordionProps = AccordionProps;
+export type PrimitiveAccordionProps = ComponentProps<typeof PrimitiveAccordion>;
 const PrimitiveAccordion: PandaComponent<typeof Accordion> = panda(Accordion);
 
-export type PrimitiveAccordionItemProps = AccordionItemProps;
+export type PrimitiveAccordionItemProps = ComponentProps<
+  typeof PrimitiveAccordionItem
+>;
 export const PrimitiveAccordionItem: PandaComponent<typeof AccordionItem> =
   panda(AccordionItem);
 
-export type PrimitiveAccordionTriggerProps = AccordionTriggerProps;
+export type PrimitiveAccordionTriggerProps = ComponentProps<
+  typeof PrimitiveAccordionTrigger
+>;
 export const PrimitiveAccordionTrigger = panda(AccordionTrigger);
 
-export type PrimitiveAccordionContentProps = AccordionContentProps;
+export type PrimitiveAccordionContentProps = ComponentProps<
+  typeof PrimitiveAccordionContent
+>;
 export const PrimitiveAccordionContent = panda(AccordionContent);
 
 export default PrimitiveAccordion;